refactor(Core.PackageIndexing): nest all PackageIndexing types; rename straggler files
### Core.PackageIndexing sub-namespace anchored

New Core.PackageIndexing.swift declares the sub-namespace. All
PackageIndexing-folder types now live under Core.PackageIndexing.

### Existing CPI files: extension Core -> extension Core.PackageIndexing

Six files were already named Core.PackageIndexing+X.swift but extended
Core directly (the namespace mismatch tracked in #371/#374). Repointed:

- Core.PackageIndexing+ManifestCache.swift
- Core.PackageIndexing+PackageArchiveExtractor.swift
- Core.PackageIndexing+PackageAvailabilityAnnotator.swift
- Core.PackageIndexing+PackageDependencyResolver.swift
- Core.PackageIndexing+PackageFileKind.swift
- Core.PackageIndexing+ResolvedPackagesStore.swift

### Stragglers wrapped + renamed

PackageFetcher (5 file-scope companion types + internal Error enum):

- PackageInfo            -> Core.PackageIndexing.PackageFetcher.PackageInfo
- PackageFetchOutput     -> Core.PackageIndexing.PackageFetcher.FetchOutput
- PackageFetchCheckpoint -> Core.PackageIndexing.PackageFetcher.Checkpoint
- PackageFetchStatistics -> Core.PackageIndexing.PackageFetcher.Statistics
- PackageFetchProgress   -> Core.PackageIndexing.PackageFetcher.Progress
- PackageFetchError      -> Core.PackageIndexing.PackageFetcher.Error

PriorityPackageGenerator (6 companions):

- PriorityPackageList    -> Core.PackageIndexing.PriorityPackageGenerator.PriorityPackageList
- PriorityLevels         -> Core.PackageIndexing.PriorityPackageGenerator.PriorityLevels
- TierInfo               -> Core.PackageIndexing.PriorityPackageGenerator.TierInfo
- PriorityPackageInfo    -> Core.PackageIndexing.PriorityPackageGenerator.PriorityPackageInfo
- PackageStats           -> Core.PackageIndexing.PriorityPackageGenerator.PackageStats
- PriorityPackageError   -> Core.PackageIndexing.PriorityPackageGenerator.Error

PriorityPackagesCatalog + 4 companions wrapped:

- PriorityPackage        -> Core.PackageIndexing.PriorityPackage
- PriorityTier           -> Core.PackageIndexing.PriorityTier
- PriorityTiers          -> Core.PackageIndexing.PriorityTiers
- PriorityPackageStats   -> Core.PackageIndexing.PriorityPackageStats

PackageDocumentationDownloader + PackageDownloadError moved into Core.PackageIndexing.

### File renames per <Namespace>+<Type>.swift convention

- PackageFetcher.swift                  -> Core.PackageIndexing+PackageFetcher.swift
- PriorityPackageGenerator.swift        -> Core.PackageIndexing+PriorityPackageGenerator.swift
- PriorityPackagesCatalog.swift         -> Core.PackageIndexing+PriorityPackagesCatalog.swift
- PackageDocumentationDownloader.swift  -> Core.PackageIndexing+PackageDocumentationDownloader.swift

New anchor: Core.PackageIndexing.swift.

### External caller sweep (production + tests)

- Search/PackageIndex.swift, PackageIndexer.swift, Search+IndexBuilder.swift
- CLI/Command+Fetch.swift, Command+Doctor.swift
- Core/Core+Crawler.swift
- TUI/Helpers/PackageActions.swift, PackageCurator.swift
- 5 CoreTests files

All Core.PackageFetcher / PriorityPackageGenerator / PriorityPackagesCatalog /
PackageDocumentationDownloader / Core.ResolvedPackagesStore / Core.PackageArchiveExtractor /
Core.ExtractedFile / Core.PackageFileKind / Core.ManifestCache /
Core.PackageDependencyResolver / Core.PackageAvailabilityAnnotator
references qualified through Core.PackageIndexing.<type>.

### Verification

- xcrun swift build (clean): 0 warnings, 0 errors
- xcrun swift test (clean):  1300/1300 pass
