+++ Packages/Sources/Core/PackageIndexing/Core.PackageIndexing+PackageArchiveExtractor.swift
@@ -3,7 +3,7 @@ import Foundation
 import SharedConstants
 import SharedCore
 
-extension Core {
+extension Core.PackageIndexing {
     /// Fetches a repo's source tarball from `codeload.github.com/<owner>/<repo>/tar.gz/<ref>`
     /// and returns its text content as `[ExtractedFile]` in memory. The caller (the
     /// indexer) is expected to consume the list directly into SQLite; nothing is
